requirements.txt
albumentations==1.3.1
braceexpand==0.1.7
detectron2==0.6
easydict==1.11
gdown==4.7.1
imageio==2.31.5
imgaug==0.4.0
joblib==1.3.2
kornia==0.7.0
matplotlib==3.8.0
memory_profiler==0.61.0
numpy==1.26.1
omegaconf==2.3.0
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
pandas==2.1.3
Pillow==10.1.0
pytorch_lightning==2.1.1
PyYAML==6.0.1
Requests==2.31.0
scikit_learn==1.3.1
scipy==1.11.3
six==1.16.0
skimage==0.0
sync_batchnorm==0.0.1
tensorflow==2.14.0
tensorflow_intel==2.14.0
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
webdataset==0.2.75
